% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgressEbar.R
\name{Progress.Ebar}
\alias{Progress.Ebar}
\title{ProgressEbar}
\usage{
Progress.Ebar(
  min = 0,
  max = 1,
  style = 3,
  char = "/",
  char.color = "cyan",
  bgchar = "/",
  bgchar.color = "red",
  text = "逃げちゃだめだ ",
  text.color = "blue"
)
}
\arguments{
\item{max}{set the max loop numbers}

\item{char}{using style1,2,3.   only half-width alphanumeric character}

\item{char.color}{"red", "yellow", "green", "blue", "violet", "cyan" or "grey"}

\item{bgchar}{using style3}

\item{text}{using style4.}
}
\description{
\code{Progress.Ebar} Setting the progress bar parameters
}
\examples{
#n<-100
#pe<-Progress.Ebar(max = n，style＝3)
}
