% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgressEber.R
\name{PW.G.labeling}
\alias{PW.G.labeling}
\title{タイトル（関数がすることを書くと良い）}
\usage{
PW.G.labeling(
  pw = NULL,
  titlelab = "title",
  xlab = "xlab",
  ylab = "ylab",
  titlefont = 25,
  xfont = 20,
  yfont = 20,
  xhigh = 0.3
)
}
\arguments{
\item{引数}{説明}
}
\value{
出力の説明
}
\description{
\code{関数名} 説明（関数がすることを書くと良い）
}
\examples{
関数の例
}
