% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgressEbar.R
\name{set.PE}
\alias{set.PE}
\title{set.PE}
\usage{
set.PE(pe, value)
}
\description{
\code{set.PE} Set up a progress bar
}
\examples{
#n<-100
#pe<-Progress.Ebar(min = 1,max = n)
#for (i in seq(n)) {
#set.PE(pe,i)
#Sys.sleep(1 / 100)
#}
}
