% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgressEber.R
\name{Progress.Ebar}
\alias{Progress.Ebar}
\title{ProgressEber}
\usage{
Progress.Ebar(
  min = 0,
  max = 1,
  style = 3,
  char = "/",
  char.color = "cyan",
  bgchar = "/",
  bgchar.color = "red",
  text = "逃げちゃだめだ",
  text.color = "blue"
)
}
\description{
\code{Progress.Eber}
}
