% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgressEber.R
\name{Progress.Ebar}
\alias{Progress.Ebar}
\title{タイトル（関数がすることを書くと良い）}
\usage{
Progress.Ebar(
  min = 0,
  max = 1,
  style = 3,
  char = "/",
  char.color = "cyan",
  bgchar = "/",
  bgchar.color = "red",
  text = "word",
  text.color = "blue"
)
}
\arguments{
\item{引数}{説明}
}
\value{
出力の説明
}
\description{
\code{関数名} 説明（関数がすることを書くと良い）
}
\examples{
関数の例
}
