% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgressEber.R
\name{PW.G.labeling}
\alias{PW.G.labeling}
\title{PW.G.labeling}
\usage{
PW.G.labeling(
  pw = NULL,
  titlelab = "title",
  xlab = "xlab",
  ylab = "ylab",
  titlefont = 25,
  xfont = 20,
  yfont = 20,
  xhigh = 0.3
)
}
\description{
\code{PW.G.labeling}
}
